showPartsEdit から Promise を削除し、 checkPartsEdit に変更
